requirements.txt
SQLAlchemy==2.0.43
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
alembic==1.16.5
SQLAlchemy-Utils==0.41.0
